the big shuffle
I organized all the variables and functions into categorized header files
